chore(correct-redirect): поправлен редирект
