+++ src/lib/auth-redirect.ts
@@ -1,0 +1,13 @@
+﻿import { notFound, redirect } from "next/navigation";
+
+import { getDefaultPath, type SessionAccessUser } from "./auth-access";
+
+export function redirectToDefaultPathOr404(user: SessionAccessUser): never {
+  const targetPath = getDefaultPath(user);
+
+  if (!targetPath) {
+    notFound();
+  }
+
+  redirect(targetPath);
+}
